Greater unity between Woodbury & SymWoodbury
There were quite a few methods defined for one and not the other, and 
the entire solution strategy was different. This centralizes much of the 
code. Some advantages and major changes:
- it's now possible to pass a factorization object as `A` to 
`SymWoodbury`
- The `SymWoodbury` constructor now checks the symmetry of `A` and `D`. 
This was not checked before and indeed many tests formerly constructed 
asymmetric SymWoodbury matrices
- arithmetic operations that were formerly specific to SymWoodbury are 
now implemented for Woodbury